Upstream changes to support Cloud Operator and the new CloudLogStorage controller
