+++ pkg/controller/logstorage/logstorage_controller.go
@@ -18,14 +18,11 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/tigera/operator/pkg/crypto"
-
 	cmnv1 "github.com/elastic/cloud-on-k8s/pkg/apis/common/v1"
 	esv1 "github.com/elastic/cloud-on-k8s/pkg/apis/elasticsearch/v1"
 	kbv1 "github.com/elastic/cloud-on-k8s/pkg/apis/kibana/v1"
-	"golang.org/x/crypto/bcrypt"
-
 	operatorv1 "github.com/tigera/operator/api/v1"
+	"github.com/tigera/operator/pkg/controller/logstorage/common"
 	"github.com/tigera/operator/pkg/controller/options"
 	"github.com/tigera/operator/pkg/controller/status"
 	"github.com/tigera/operator/pkg/controller/utils"
@@ -60,15 +57,8 @@ import (
 var log = logf.Log.WithName("controller_logstorage")
 
 const (
-	tigeraElasticsearchUserSecretLabel = "tigera-elasticsearch-user"
-	defaultElasticsearchShards         = 1
-	defaultEckOperatorMemorySetting    = "512Mi"
-	DefaultElasticsearchStorageClass   = "tigera-elasticsearch"
-
-	// Mark any secret containing credentials for ES gateway with this label key/value. This will allow ES gateway to watch only the
-	// releveant secrets it needs.
-	ESGatewaySelectorLabel      = "esgateway.tigera.io/secrets"
-	ESGatewaySelectorLabelValue = "credentials"
+	defaultEckOperatorMemorySetting  = "512Mi"
+	DefaultElasticsearchStorageClass = "tigera-elasticsearch"
 )
 
 // Add creates a new LogStorage Controller and adds it to the Manager. The Manager will set fields on the Controller
@@ -157,15 +147,15 @@ func add(mgr manager.Manager, r reconcile.Reconciler) error {
 	// the utils folder where the other watch logic is.
 	err = c.Watch(&source.Kind{Type: &corev1.Secret{}}, &handler.EnqueueRequestForObject{}, &predicate.Funcs{
 		CreateFunc: func(e event.CreateEvent) bool {
-			_, hasLabel := e.Object.GetLabels()[tigeraElasticsearchUserSecretLabel]
+			_, hasLabel := e.Object.GetLabels()[common.TigeraElasticsearchUserSecretLabel]
 			return e.Object.GetNamespace() == rmeta.OperatorNamespace() && hasLabel
 		},
 		UpdateFunc: func(e event.UpdateEvent) bool {
-			_, hasLabel := e.ObjectNew.GetLabels()[tigeraElasticsearchUserSecretLabel]
+			_, hasLabel := e.ObjectNew.GetLabels()[common.TigeraElasticsearchUserSecretLabel]
 			return e.ObjectNew.GetNamespace() == rmeta.OperatorNamespace() && hasLabel
 		},
 		DeleteFunc: func(e event.DeleteEvent) bool {
-			_, hasLabel := e.Object.GetLabels()[tigeraElasticsearchUserSecretLabel]
+			_, hasLabel := e.Object.GetLabels()[common.TigeraElasticsearchUserSecretLabel]
 			return e.Object.GetNamespace() == rmeta.OperatorNamespace() && hasLabel
 		},
 	})
@@ -434,8 +424,8 @@ func (r *ReconcileLogStorage) Reconcile(ctx context.Context, request reconcile.R
 	customerProvidedCert := false
 
 	if managementClusterConnection == nil {
-		var flowShards = calculateFlowShards(ls.Spec.Nodes, defaultElasticsearchShards)
-		clusterConfig = relasticsearch.NewClusterConfig(render.DefaultElasticsearchClusterName, ls.Replicas(), defaultElasticsearchShards, flowShards)
+		var flowShards = common.CalculateFlowShards(ls.Spec.Nodes, common.DefaultElasticsearchShards)
+		clusterConfig = relasticsearch.NewClusterConfig(render.DefaultElasticsearchClusterName, ls.Replicas(), common.DefaultElasticsearchShards, flowShards)
 
 		// Check if there is a StorageClass available to run Elasticsearch on.
 		if err := r.client.Get(ctx, client.ObjectKey{Name: ls.Spec.StorageClassName}, &storagev1.StorageClass{}); err != nil {
@@ -451,7 +441,7 @@ func (r *ReconcileLogStorage) Reconcile(ctx context.Context, request reconcile.R
 			return reconcile.Result{}, nil
 		}
 
-		if gatewayCertSecret, publicCertSecret, customerProvidedCert, err = r.getESGatewayCertificateSecrets(ctx, install); err != nil {
+		if gatewayCertSecret, publicCertSecret, customerProvidedCert, err = common.GetESGatewayCertificateSecrets(ctx, install, r.client, r.clusterDomain, log); err != nil {
 			reqLogger.Error(err, err.Error())
 			r.status.SetDegraded("Failed to create Elasticsearch Gateway secrets", err.Error())
 			return reconcile.Result{}, err
@@ -598,7 +588,7 @@ func (r *ReconcileLogStorage) Reconcile(ctx context.Context, request reconcile.R
 			return reconcile.Result{}, err
 		}
 
-		kubeControllersGatewaySecret, kubeControllersVerificationSecret, kubeControllersSecureUserSecret, err := r.createKubeControllersSecrets(ctx, esAdminUserSecret)
+		kubeControllersGatewaySecret, kubeControllersVerificationSecret, kubeControllersSecureUserSecret, err := common.CreateKubeControllersSecrets(ctx, esAdminUserSecret, r.client)
 		if err != nil {
 			reqLogger.Error(err, err.Error())
 			r.status.SetDegraded("Failed to create kube-controllers secrets for Elasticsearch gateway", "")
@@ -711,162 +701,6 @@ func (r *ReconcileLogStorage) Reconcile(ctx context.Context, request reconcile.R
 	return reconcile.Result{}, nil
 }
 
-// Deletes the given ECK managed cert secret.
-func (r *ReconcileLogStorage) deleteInvalidECKManagedPublicCertSecret(ctx context.Context, secret *corev1.Secret) error {
-	log.Info(fmt.Sprintf("Deleting invalid cert secret %q in %q namespace", secret.Name, secret.Namespace))
-	return r.client.Delete(ctx, secret)
-}
-
-// createKubeControllersSecrets checks for the existence of the secrets necessary for Kube controllers to access Elasticsearch through ES gateway and
-// creates them if they are missing. Kube controllers no longer uses admin credentials to make requests directly to Elasticsearch. Instead, gateway credentials
-// are generated and stored in the user secret, a hashed version of the credentials is stored in the tigera-elasticsearch namespace for ES Gateway to retrieve and use to compare
-// the gateway credentials, and a secret containing real admin level credentials is created and stored in the tigera-elasticsearch namespace to be swapped in once
-// ES Gateway has confirmed that the gateway credentials match.
-func (r *ReconcileLogStorage) createKubeControllersSecrets(ctx context.Context, esAdminUserSecret *corev1.Secret) (*corev1.Secret, *corev1.Secret, *corev1.Secret, error) {
-	kubeControllersGatewaySecret, err := utils.GetSecret(ctx, r.client, render.ElasticsearchKubeControllersUserSecret, rmeta.OperatorNamespace())
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	if kubeControllersGatewaySecret == nil {
-		password := crypto.GeneratePassword(16)
-		kubeControllersGatewaySecret = &corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      render.ElasticsearchKubeControllersUserSecret,
-				Namespace: rmeta.OperatorNamespace(),
-			},
-			Data: map[string][]byte{
-				"username": []byte(render.ElasticsearchKubeControllersUserName),
-				"password": []byte(password),
-			},
-		}
-	}
-	hashedPassword, err := bcrypt.GenerateFromPassword(kubeControllersGatewaySecret.Data["password"], bcrypt.MinCost)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	kubeControllersVerificationSecret, err := utils.GetSecret(ctx, r.client, render.ElasticsearchKubeControllersVerificationUserSecret, render.ElasticsearchNamespace)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	if kubeControllersVerificationSecret == nil {
-		kubeControllersVerificationSecret = &corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      render.ElasticsearchKubeControllersVerificationUserSecret,
-				Namespace: render.ElasticsearchNamespace,
-				Labels: map[string]string{
-					ESGatewaySelectorLabel: ESGatewaySelectorLabelValue,
-				},
-			},
-			Data: map[string][]byte{
-				"username": []byte(render.ElasticsearchKubeControllersUserName),
-				"password": hashedPassword,
-			},
-		}
-	}
-
-	kubeControllersSecureUserSecret, err := utils.GetSecret(ctx, r.client, render.ElasticsearchKubeControllersSecureUserSecret, render.ElasticsearchNamespace)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	if kubeControllersSecureUserSecret == nil {
-		kubeControllersSecureUserSecret = &corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      render.ElasticsearchKubeControllersSecureUserSecret,
-				Namespace: render.ElasticsearchNamespace,
-				Labels: map[string]string{
-					ESGatewaySelectorLabel: ESGatewaySelectorLabelValue,
-				},
-			},
-			Data: map[string][]byte{
-				"username": []byte("elastic"),
-				"password": esAdminUserSecret.Data["elastic"],
-			},
-		}
-	}
-
-	return kubeControllersGatewaySecret, kubeControllersVerificationSecret, kubeControllersSecureUserSecret, nil
-}
-
-// getESGatewayCertificateSecrets retrieves certificate secrets needed for ES Gateway to run or for
-// components to communicate with Elasticsearch/Kibana through ES Gateway. The order of the secrets returned are:
-// 1) The certificate/key secret to be mounted by ES Gateway and used to authenticate requests before
-// proxying to Elasticsearch/Kibana (in the operator namespace). If the user didn't create this secret, it is created.
-// 2) The certificate mounted by other clients that connect to Elasticsearch/Kibana through ES Gateway (in the operator namespace).
-// The final return value is used to indicate that the certificate secret was provided by the customer. This
-// ensures that we do not re-render the secret in the Operator Namespace and overwrite the OwnerReference.
-func (r *ReconcileLogStorage) getESGatewayCertificateSecrets(ctx context.Context, instl *operatorv1.InstallationSpec) (*corev1.Secret, *corev1.Secret, bool, error) {
-	var publicCertSecret *corev1.Secret
-
-	svcDNSNames := dns.GetServiceDNSNames(render.ElasticsearchServiceName, render.ElasticsearchNamespace, r.clusterDomain)
-	svcDNSNames = append(svcDNSNames, dns.GetServiceDNSNames(esgateway.ServiceName, render.ElasticsearchNamespace, r.clusterDomain)...)
-
-	// Get the secret - might be nil
-	oprKeyCert, err := utils.GetSecret(ctx, r.client, render.TigeraElasticsearchCertSecret, rmeta.OperatorNamespace())
-	if err != nil {
-		return nil, nil, false, err
-	}
-
-	// Ensure that cert is valid.
-	oprKeyCert, err = utils.EnsureCertificateSecret(render.TigeraElasticsearchCertSecret, oprKeyCert, corev1.TLSPrivateKeyKey, corev1.TLSCertKey, rmeta.DefaultCertificateDuration, svcDNSNames...)
-	if err != nil {
-		return nil, nil, false, err
-	}
-
-	// Three different certificate issuers are possible:
-	// - The operator self-signed certificate
-	// - A user's BYO keypair for Elastic (uncommon)
-	// - The issuer that is provided through the certificate management feature.
-	keyCertIssuer, err := utils.GetCertificateIssuer(oprKeyCert.Data[corev1.TLSCertKey])
-	if err != nil {
-		return nil, nil, false, err
-	}
-	customerProvidedCert := !utils.IsOperatorIssued(keyCertIssuer)
-
-	// If Certificate management is enabled, we only want to trust the CA cert and let the init container handle private key generation.
-	if instl.CertificateManagement != nil {
-		cmCa := instl.CertificateManagement.CACert
-		cmIssuer, err := utils.GetCertificateIssuer(cmCa)
-		if err != nil {
-			return nil, nil, false, err
-		}
-
-		// If the issuer of the current secret is not the same as the certificate management issuer and also is not
-		// issued by the tigera-operator, it means that it is added to this cluster by the customer. This is not supported
-		// in combination with certificate management.
-		if customerProvidedCert && cmIssuer != keyCertIssuer {
-			return nil, nil, false, fmt.Errorf("certificate management does not support custom Elasticsearch secrets, please delete secret %s/%s or disable certificate management", oprKeyCert.Namespace, oprKeyCert.Name)
-		}
-
-		oprKeyCert.Data[corev1.TLSCertKey] = instl.CertificateManagement.CACert
-		publicCertSecret = render.CreateCertificateSecret(instl.CertificateManagement.CACert, relasticsearch.PublicCertSecret, rmeta.OperatorNamespace())
-	} else {
-		// Get the es gateway pub secret - might be nil
-		publicCertSecret, err = utils.GetSecret(ctx, r.client, relasticsearch.PublicCertSecret, rmeta.OperatorNamespace())
-		if err != nil {
-			return nil, nil, false, err
-		}
-
-		if publicCertSecret != nil {
-			// If the provided certificate secret (secret) is managed by the operator we need to check if the secret has the expected DNS names.
-			// If it doesn't, delete the public secret so it can get recreated.
-			if !customerProvidedCert {
-				err = utils.SecretHasExpectedDNSNames(publicCertSecret, corev1.TLSCertKey, svcDNSNames)
-				if err == utils.ErrInvalidCertDNSNames {
-					if err := r.deleteInvalidECKManagedPublicCertSecret(ctx, publicCertSecret); err != nil {
-						return nil, nil, false, err
-					}
-					publicCertSecret = render.CreateCertificateSecret(oprKeyCert.Data[corev1.TLSCertKey], relasticsearch.PublicCertSecret, rmeta.OperatorNamespace())
-				}
-			}
-		} else {
-			publicCertSecret = render.CreateCertificateSecret(oprKeyCert.Data[corev1.TLSCertKey], relasticsearch.PublicCertSecret, rmeta.OperatorNamespace())
-		}
-	}
-
-	return oprKeyCert, publicCertSecret, customerProvidedCert, nil
-}
-
 // getElasticsearchCertificateSecrets retrieves Elasticsearch certificate secrets needed for Elasticsearch to run or for
 // ES gateway to communicate with Elasticsearch. The order of the secrets returned are:
 // 1) The certificate secret needed for Elasticsearch (in the Elasticsearch namespace). If the user didn't create this it is
@@ -908,7 +742,7 @@ func (r *ReconcileLogStorage) getElasticsearchCertificateSecrets(ctx context.Con
 			// public secret so it can get recreated.
 			err = utils.SecretHasExpectedDNSNames(internalSecret, corev1.TLSCertKey, svcDNSNames)
 			if err == utils.ErrInvalidCertDNSNames {
-				if err := r.deleteInvalidECKManagedPublicCertSecret(ctx, internalSecret); err != nil {
+				if err := common.DeleteInvalidECKManagedPublicCertSecret(ctx, internalSecret, r.client, log); err != nil {
 					return nil, nil, err
 				}
 			}
@@ -967,7 +801,7 @@ func (r *ReconcileLogStorage) kibanaInternalSecrets(ctx context.Context, instl *
 	if utils.IsOperatorIssued(issuer) {
 		err = utils.SecretHasExpectedDNSNames(internalSecret, corev1.TLSCertKey, svcDNSNames)
 		if err == utils.ErrInvalidCertDNSNames {
-			if err := r.deleteInvalidECKManagedPublicCertSecret(ctx, internalSecret); err != nil {
+			if err := common.DeleteInvalidECKManagedPublicCertSecret(ctx, internalSecret, r.client, log); err != nil {
 				return nil, err
 			}
 		}
@@ -1035,20 +869,4 @@ func (r *ReconcileLogStorage) checkOIDCUsersEsResource(ctx context.Context) erro
 		return err
 	}
 	return nil
-}
-
-func calculateFlowShards(nodesSpecifications *operatorv1.Nodes, defaultShards int) int {
-	if nodesSpecifications == nil || nodesSpecifications.ResourceRequirements == nil || nodesSpecifications.ResourceRequirements.Requests == nil {
-		return defaultShards
-	}
-
-	var nodes = nodesSpecifications.Count
-	var cores, _ = nodesSpecifications.ResourceRequirements.Requests.Cpu().AsInt64()
-	var shardPerNode = int(cores) / 4
-
-	if nodes <= 0 || shardPerNode <= 0 {
-		return defaultShards
-	}
-
-	return int(nodes) * shardPerNode
 }
