+++ pkg/render/monitor.go
@@ -98,7 +98,7 @@ func (mc *monitorComponent) SupportedOSType() rmeta.OSType {
 
 func (mc *monitorComponent) Objects() ([]client.Object, []client.Object) {
 	toCreate := []client.Object{
-		createNamespace(common.TigeraPrometheusNamespace, mc.installation.KubernetesProvider),
+		CreateNamespace(common.TigeraPrometheusNamespace, mc.installation.KubernetesProvider),
 	}
 	toCreate = append(toCreate, secret.ToRuntimeObjects(secret.CopyToNamespace(common.TigeraPrometheusNamespace, mc.pullSecrets...)...)...)
 	toCreate = append(toCreate,
